renamed previous school-assignments table to avoid conflict
