+++ db/tables/semesters.ts
@@ -10,7 +10,7 @@ import { applicationsTable } from "./applications";
 import { departmentsTable } from "./departments";
 import { meetingsTable } from "./meetings";
 import { mainSchema } from "./schema";
-import { schoolAssignmentsTable } from "./school-assignments";
+import { schoolSemesterAssistantsTable } from "./school-semester-assistant";
 import { teamSemesterUsersTable } from "./team-semester-user";
 
 export const semestersTable = mainSchema.table("semesters", {
@@ -43,7 +43,7 @@ export const semestersRelations = relations(
 			fields: [semestersTable.id],
 			references: [semestersTable.lastSemesterId],
 		}),
-		schoolAssistants: many(schoolAssignmentsTable),
+		schoolAssistants: many(schoolSemesterAssistantsTable),
 		teamUsers: many(teamSemesterUsersTable),
 		applications: many(applicationsTable),
 		meetings: many(meetingsTable),
